requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.3.6
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2022.1
cron-descriptor==1.4.0
cryptography==41.0.1
defusedxml==0.7.1
discord==2.3.0
Django==4.2.2
django-allauth==0.54.0
django-appconf==1.0.5
django-bootstrap4==23.1
django-celery-beat==2.5.0
django-cleanup==8.0.0
django-crispy-forms==2.0
django-etc==1.4.0
django-hitcount==1.3.5
django-imagekit==4.1.0
django-model-cleanup==0.1.1
django-recaptcha==3.0.0
django-redis==5.4.0
django-rest-framework==0.1.0
django-rest-framework-apikeys==0.1
django-timezone-field==5.1
djangorestframework==3.14.0
ffmpeg-python==0.2.0
frozenlist==1.3.3
future==0.18.3
idna==3.4
imageio-ffmpeg==0.4.8
kombu==5.3.4
multidict==6.0.4
notifications==0.3.2
oauthlib==3.2.2
pathlib==1.0.1
pilkit==2.0
Pillow==9.5.0
prompt-toolkit==3.0.39
py-cord==2.4.1
pycparser==2.21
PyJWT==2.7.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.3
vine==5.1.0
wcwidth==0.2.6
yarl==1.9.2
